+++ placement/placement_test.go
@@ -22,12 +22,13 @@ import (
 	"testing"
 	"time"
 
-	"github.com/facebookgo/clock"
-	"github.com/golang/protobuf/proto"
 	"github.com/m3db/m3cluster/kv"
 	"github.com/m3db/m3storage/generated/proto/schema"
 	"github.com/m3db/m3x/log"
 	"github.com/m3db/m3x/time"
+
+	"github.com/facebookgo/clock"
+	"github.com/golang/protobuf/proto"
 	"github.com/stretchr/testify/require"
 )
